json_proto_conversion.cpp
